Fixed uniqueness validator with STI
Consider a class `Base` and a class `Child < Base` with a uniqueness check in `Base.some_property`

The uniqueness check did a `target.model.first(:some_property => new_value)` where `target.model` would resolve to `Child`. But nobody said that the uniqueness is scoped by the type (that should be done with `:scope => :type`).